feat: handle basic CRUD, improve user flow

+++ recipebot/adapters/repositories/sql/base/queries/create_recipe_tags_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS recipe_tag (
     id SERIAL PRIMARY KEY,
     name VARCHAR(50) NOT NULL,
 
-    group_id INTEGER REFERENCES groups (tg_chat_id),
-    user_id INTEGER REFERENCES users (tg_id)
+    group_id BIGINT REFERENCES groups (tg_chat_id),
+    user_id BIGINT REFERENCES users (tg_id)
 
 );
