+++ recipebot/adapters/repositories/sql/base/queries/create_user_groups_junction_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users_groups(
-    user_id INTEGER REFERENCES users(tg_id),
-    group_id INTEGER REFERENCES groups(tg_chat_id),
+    user_id BIGINT REFERENCES users(tg_id),
+    group_id BIGINT REFERENCES groups(tg_chat_id),
     PRIMARY KEY(user_id, group_id)
 );
